fix(aci): add check-in timeline legend to cron detector details
